Capture multiple event triggers as long as they relate to different user
This is my use case: in the Stamp collection module, the teacher can
give stamps to multiple users at once. Each time the stamp is given, the
event is triggered. The current check caused only the first processed
student received the XP, all other were ignored (as the same event was
triggered in the same context at the almost same time).

This patch allows to capture the same event in the same context if the
event's user is different. It will still prevent users themselves to
fake the xp capturing by quick reloading the browser.