Prevent the notice from being announced to screen readers. Fixes #99
Some of the version of Moodle 3.6 and 3.7 were affected by this
where the notices where given focus and caused a page scroll.
This behaviour was changed in MDL-66247 for 3.6.5, 3.7.1 and 3.8,
however as most notices of Level up! informative and not 'alerts'
as such, we decided to stop announcing them by default.

We will add support for announcing them whenever important notices
are being used in Level up!.