Neighboured leaderboard returns actual position and count
Prior to this patch, the position and count were assumed based
on the number of neighbours in the leaderboard. This lead to complex
handling of offsets when retrieving the ranking, and also created
other issues if the leaderboard were to be paginated.

The methods now return the actual position and count relative
to the user the leaderboard is relative to. So when the user
is appears at the top of its neighbours ranking, its position
will now be 0.