Suppressing notice due to missing return type. Fixes #150
The plugin being compatible from 3.3 to 4.1, we are expected to
support PHP 5.6, although we should probably stop doing so. Either
way, the expected type 'mixed' cannot be used in PHP 7, so we're
better off silencing this until further notice. The type 'array'
could apparently be used, but we'll make the switch when we
officially drop support for PHP 5.6, or require Moodle 3.4.

References:

- https://3v4l.org/6B9OR
- https://3v4l.org/ufnAt
- https://3v4l.org/udNW1