Address randomly failing test due to dependency resolution order
It had been a while since we knew that the $forwholesite variable could
be out of sync with the admin config, which can be seen in tests.
However, the weird one here is that this test reliably passed on
many versions (Moodle, PHP) with Postgres.

Always relying on the admin config makes it more robust.